backend/pyproject.toml
[project]
name = "kaasu-backend"
version = "0.1.0"
description = "Kaasu backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.95.0",
	"uvicorn[standard]>=0.22.0",
	"sqlalchemy>=2.0.0",
	"psycopg2-binary>=2.9.0",
	"python-dotenv>=1.0.0",
]

[tool.uv]
